src/requirements.txt
beautifulsoup4 
python-dotenv
 
chromadb

transformers>=4.35.0 
peft  
accelerate 
bitsandbytes 
safetensors 
sentencepiece 

chromadb 
langchain-community
langchain 
gradio>=4.0.0 
pypdf
sentence_transformers>=4.35.0 
sentence-transformers>=4.35.0 
huggingface_hub